pyproject.toml
[tool.poetry]
name = "rock_paper_scissors"
version = "0.1.0"
description = "'Rock, Paper, Scissors' with FastAPI and WebSockets"
authors = ["Artsiom Zhvanko <azhvanko@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
uvicorn = {extras = ["uvloop"], version = "^0.17.6"}
uvloop = "^0.16.0"
websockets = "^10.3"
python-dotenv = "^0.20.0"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.37"}
asyncpg = "^0.25.0"

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.18.3"
requests = "^2.28.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
mock_use_standalone_module = "true"
addopts = "-v --disable-pytest-warnings --full-trace"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
